backend/requirements.txt
fastapi
uvicorn
sqlalchemy
scikit-learn
pandas
python-multipart
pydantic
passlib[bcrypt]
python-jose[cryptography]
websockets
